mirrors: Move tests to mirrors/tests
Move the tests to separate files in mirrors/tests and expand the model
tests with tests for the Mirror class.
